# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  - url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
  /api/login:
    post:
      summary: Login
      operationId: login
      description: "Authenticate using email or phone with a password. Returns a Sanctum token and user data.\nWorks for both admin users and customers."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 1
                    name: John
                    email: user@example.com
                    role: customer
                  token: 1|abc123...
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: John
                      email:
                        type: string
                        example: user@example.com
                      role:
                        type: string
                        example: customer
                    description: 'The authenticated user or customer object.'
                  token:
                    type: string
                    example: 1|abc123...
                    description: 'The Sanctum plain-text token.'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid credentials.'
                properties:
                  message:
                    type: string
                    example: 'Invalid credentials.'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Identifier (email or phone) is required.'
                properties:
                  message:
                    type: string
                    example: 'Identifier (email or phone) is required.'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: 'The account password.'
                  example: secret123
                email:
                  type: string
                  description: 'The email address (required if phone is not provided).'
                  example: user@example.com
                phone:
                  type: string
                  description: 'The phone number (required if email is not provided).'
                  example: '07701234567'
              required:
                - password
      security: []
...
  /api/catalog:
    get:
      summary: 'List catalog items'
      operationId: listCatalogItems
      description: 'Retrieve a paginated list of games and/or programs with filtering options.'
      parameters:
        -
          in: query
          name: per_page
          description: 'Number of items per page (max 100).'
          example: 15
          required: false
          schema:
            type: integer
            description: 'Number of items per page (max 100).'
            example: 15
        -
          in: query
          name: page
          description: 'Page number.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Page number.'
            example: 1
        -
          in: query
          name: type
          description: "Filter by type: 'all', 'game', or 'program'."
          example: all
          required: false
          schema:
            type: string
            description: "Filter by type: 'all', 'game', or 'program'."
            example: all
        -
          in: query
          name: search
          description: 'Search by name or tags.'
          example: action
          required: false
          schema:
            type: string
            description: 'Search by name or tags.'
            example: action
        -
          in: query
          name: category_id
          description: 'Filter by category ID.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'Filter by category ID.'
            example: 1
        -
          in: query
          name: platform_id
          description: 'Filter by platform ID.'
          example: 2
          required: false
          schema:
            type: integer
            description: 'Filter by platform ID.'
            example: 2
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name_en: 'Game Name'
                      name_ar: 'اسم اللعبة'
                      type: game
                      category:
                        id: 1
                        name_en: Action
                      platform:
                        id: 1
                        name_en: PlayStation
                  meta:
                    current_page: 1
                    last_page: 5
                    per_page: 15
                    total: 75
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name_en: 'Game Name'
                        name_ar: 'اسم اللعبة'
                        type: game
                        category:
                          id: 1
                          name_en: Action
                        platform:
                          id: 1
                          name_en: PlayStation
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name_en:
                          type: string
                          example: 'Game Name'
                        name_ar:
                          type: string
                          example: 'اسم اللعبة'
                        type:
                          type: string
                          example: game
                        category:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            name_en:
                              type: string
                              example: Action
                        platform:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            name_en:
                              type: string
                              example: PlayStation
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 5
                      per_page:
                        type: integer
                        example: 15
                      total:
                        type: integer
                        example: 75
      tags:
        - Catalog
      security: []
  /api/categories:
    get:
      summary: 'List all categories'
      operationId: listAllCategories
      description: 'Retrieve a list of all available categories with bilingual names.'
      parameters: []
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    name_en:
                      type: string
                      example: Action
                    name_ar:
                      type: string
                      example: أكشن
                    description_en:
                      type: string
                      example: 'Action games'
                    description_ar:
                      type: string
                      example: 'ألعاب أكشن'
                    category_type_id:
                      type: integer
                      example: 1
                example:
                  -
                    id: 1
                    name_en: Action
                    name_ar: أكشن
                    description_en: 'Action games'
                    description_ar: 'ألعاب أكشن'
                    category_type_id: 1
                  -
                    id: 2
                    name_en: Sports
                    name_ar: رياضة
                    description_en: 'Sports games'
                    description_ar: 'ألعاب رياضية'
                    category_type_id: 1
      tags:
        - Catalog
      security: []
    post:
      summary: ''
      operationId: postApiCategories
      description: ''
      parameters: []
      responses: {}
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category_type_id:
                  type: string
                  description: 'Must match an existing stored value.'
                  example: architecto
                name_en:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                name_ar:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: g
                description_en:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                description_ar:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
              required:
                - category_type_id
                - name_en
                - name_ar
      security: []
...
  /api/customers:
    post:
      summary: 'Register a customer'
      operationId: registerACustomer
      description: 'Create a new customer account. Public endpoint.'
      parameters: []
      responses:
        201:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: '{"id": 1, "name": "John Doe", "phone": "07701234567", "governorate": {...}}'
      tags:
        - Customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Full name.'
                  example: 'John Doe'
                phone:
                  type: string
                  description: 'Iraqi phone number (must be unique).'
                  example: '07701234567'
                email:
                  type: string
                  description: 'Email address (must be unique).'
                  example: john@example.com
                  nullable: true
                password:
                  type: string
                  description: 'Minimum 8 characters, no spaces.'
                  example: secret123
                governorate_id:
                  type: integer
                  description: 'Governorate ID.'
                  example: 1
                address:
                  type: string
                  description: 'Address details.'
                  example: architecto
                  nullable: true
                nearest_service_point:
                  type: string
                  description: 'Nearest service point.'
                  example: architecto
                  nullable: true
              required:
                - name
                - phone
                - password
                - governorate_id
      security: []
    get:
      summary: 'List customers'
      operationId: listCustomers
      description: 'Get all registered customers with their governorates. Admin only.'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Customers
  '/api/customers/{customer_id}':
    get:
      summary: 'Show a customer'
      operationId: showACustomer
      description: 'Get details of a single customer. Admin only.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: '{"id": 1, "name": "John Doe", "phone": "07701234567", "email": "john@example.com", "governorate": {...}}'
      tags:
        - Customers
    put:
      summary: 'Update a customer'
      operationId: updateACustomer
      description: "Update an existing customer's information. Admin only."
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: '{"id": 1, "name": "John Doe", "phone": "07701234567", "governorate": {...}}'
      tags:
        - Customers
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Full name.'
                  example: architecto
                phone:
                  type: string
                  description: 'Iraqi phone number.'
                  example: architecto
                email:
                  type: string
                  description: 'Email address.'
                  example: gbailey@example.net
                governorate_id:
                  type: integer
                  description: 'Governorate ID.'
                  example: 16
                address:
                  type: string
                  description: 'Address details.'
                  example: architecto
                  nullable: true
                nearest_service_point:
                  type: string
                  description: 'Nearest service point.'
                  example: architecto
                  nullable: true
    delete:
      summary: 'Delete a customer'
      operationId: deleteACustomer
      description: 'Remove a customer account. Admin only.'
      parameters: []
      responses:
        204:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: ''
      tags:
        - Customers
    parameters:
      -
        in: path
        name: customer_id
        description: 'The ID of the customer.'
        example: 1
        required: true
        schema:
          type: integer
      -
        in: path
        name: customer
        description: 'The customer ID.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/programs/{id}':
    get:
      summary: ''
      operationId: getApiProgramsId
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name_en: 'Adobe Photoshop 2025'
                  description_en: 'Adobe Photoshop 2025 is a professional Photo Editing application available on Windows.'
                  name_ar: 'أدوبي فوتوشوب 2025'
                  description_ar: 'أدوبي فوتوشوب 2025 هو تطبيق Photo Editing احترافي متوفر على Windows.'
                  category_id: 23
                  platform_id: 18
                  tags:
                    - design
                    - photo-editing
                    - graphics
                  size_gb: 3.8
                  downloads: 3000000
                  date_release: '2025-05-20T00:00:00.000000Z'
                  created_at: '2026-06-27T12:04:46.000000Z'
                  updated_at: '2026-06-27T12:04:46.000000Z'
                  category:
                    id: 23
                    category_type_id: 2
                    name_en: 'Photo Editing'
                    name_ar: 'تحرير صور'
                    description_en: 'Image editing and graphic design software.'
                    description_ar: 'برامج تحرير الصور والتصميم الجرافيكي.'
                    created_at: null
                    updated_at: null
                  platform:
                    id: 18
                    name_en: Windows
                    name_ar: ويندوز
                    description_en: 'Microsoft Windows operating system.'
                    description_ar: 'نظام تشغيل ويندوز من مايكروسوفت.'
                    created_at: null
                    updated_at: null
                properties:
                  id:
                    type: integer
                    example: 1
                  name_en:
                    type: string
                    example: 'Adobe Photoshop 2025'
                  description_en:
                    type: string
                    example: 'Adobe Photoshop 2025 is a professional Photo Editing application available on Windows.'
                  name_ar:
                    type: string
                    example: 'أدوبي فوتوشوب 2025'
                  description_ar:
                    type: string
                    example: 'أدوبي فوتوشوب 2025 هو تطبيق Photo Editing احترافي متوفر على Windows.'
                  category_id:
                    type: integer
                    example: 23
                  platform_id:
                    type: integer
                    example: 18
                  tags:
                    type: array
                    example:
                      - design
                      - photo-editing
                      - graphics
                    items:
                      type: string
                  size_gb:
                    type: number
                    example: 3.8
                  downloads:
                    type: integer
                    example: 3000000
                  date_release:
                    type: string
                    example: '2025-05-20T00:00:00.000000Z'
                  created_at:
                    type: string
                    example: '2026-06-27T12:04:46.000000Z'
                  updated_at:
                    type: string
                    example: '2026-06-27T12:04:46.000000Z'
                  category:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 23
                      category_type_id:
                        type: integer
                        example: 2
                      name_en:
                        type: string
                        example: 'Photo Editing'
                      name_ar:
                        type: string
                        example: 'تحرير صور'
                      description_en:
                        type: string
                        example: 'Image editing and graphic design software.'
                      description_ar:
                        type: string
                        example: 'برامج تحرير الصور والتصميم الجرافيكي.'
                      created_at:
                        type: string
                        example: null
                        nullable: true
                      updated_at:
                        type: string
                        example: null
                        nullable: true
                  platform:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 18
                      name_en:
                        type: string
                        example: Windows
                      name_ar:
                        type: string
                        example: ويندوز
                      description_en:
                        type: string
                        example: 'Microsoft Windows operating system.'
                      description_ar:
                        type: string
                        example: 'نظام تشغيل ويندوز من مايكروسوفت.'
                      created_at:
                        type: string
                        example: null
                        nullable: true
                      updated_at:
                        type: string
                        example: null
                        nullable: true
      tags:
        - Endpoints
      security: []
    put:
      summary: ''
      operationId: putApiProgramsId
      description: ''
      parameters: []
      responses: {}
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name_en:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                name_ar:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                description_en:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                description_ar:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                category_id:
                  type: string
                  description: 'Must match an existing stored value.'
                  example: null
                platform_id:
                  type: string
                  description: 'Must match an existing stored value.'
                  example: null
                tags:
                  type: array
                  description: ''
                  example:
                    - architecto
                  items:
                    type: string
                size_gb:
                  type: number
                  description: ''
                  example: 4326.41688
                downloads:
                  type: integer
                  description: 'Must be at least 0.'
                  example: 77
                  nullable: true
                date_release:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2026-06-27T12:09:17'
                  nullable: true
      security: []
    delete:
      summary: ''
      operationId: deleteApiProgramsId
      description: ''
      parameters: []
      responses: {}
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the program.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/packages/{id}':
    get:
      summary: 'Show a package'
      operationId: showAPackage
      description: 'Get details of a single package by ID with its platform, category type, and items.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: '{"id": 1, "name_en": "Gaming Pack 1", "name_ar": "حزمة ألعاب ١", "platform": {...}, "package_category_type": {...}, "items": [...]}'
      tags:
        - Packages
      security: []
    put:
      summary: 'Update a package'
      operationId: updateAPackage
      description: 'Update an existing package. Admin only.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: '{"id": 1, "name_en": "Gaming Pack 1", "name_ar": "حزمة ألعاب ١", ...}'
      tags:
        - Packages
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name_en:
                  type: string
                  description: 'English name.'
                  example: architecto
                name_ar:
                  type: string
                  description: 'Arabic name.'
                  example: architecto
                description_en:
                  type: string
                  description: 'English description.'
                  example: architecto
                  nullable: true
                description_ar:
                  type: string
                  description: 'Arabic description.'
                  example: architecto
                  nullable: true
                platform_id:
                  type: integer
                  description: 'Platform ID.'
                  example: 16
                package_category_type_id:
                  type: integer
                  description: 'Package category type ID.'
                  example: 16
    delete:
      summary: 'Delete a package'
      operationId: deleteAPackage
      description: 'Remove a package. Admin only.'
      parameters: []
      responses:
        204:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: ''
      tags:
        - Packages
    parameters:
      -
        in: path
        name: id
        description: 'The package ID.'
        example: 1
        required: true
        schema:
          type: integer
...